+++ src/main/scala/org/constellation/ConstellationNode.scala
@@ -4,7 +4,7 @@ import java.net.InetSocketAddress
 import java.security.KeyPair
 
 import better.files.File
-import cats.effect.{Clock, ContextShift, ExitCode, IO, IOApp, Resource, Sync, Timer}
+import cats.effect.{Clock, ContextShift, ExitCode, IO, IOApp, Resource, Sync, SyncIO, Timer}
 import cats.syntax.all._
 import com.typesafe.config.{Config, ConfigFactory}
 import com.typesafe.scalalogging.StrictLogging
@@ -47,16 +47,7 @@ import scala.util.Try
 /**
   * Main entry point for starting a node
   */
-object ConstellationNode extends IOApp {
-
-  def run(args: List[String]): IO[ExitCode] =
-    IO.shift(ConstellationExecutionContext.unbounded) >> ConstellationNodeUnbounded.run(args)
-}
-
-object ConstellationNodeUnbounded {
-  implicit val cs: ContextShift[IO] = IO.contextShift(ConstellationExecutionContext.unbounded)
-  implicit val timer: Timer[IO] = IO.timer(ConstellationExecutionContext.unbounded)
-
+object ConstellationNode extends IOApp with IOApp.WithContext {
   implicit val logger: Logger[IO] = Slf4jLogger.getLogger[IO]
 
   final val LocalConfigFile = "local_config"
@@ -65,6 +56,11 @@ object ConstellationNodeUnbounded {
   implicit val clock: Clock[IO] = Clock.create[IO]
 
   import constellation._
+
+  override protected def executionContextResource: Resource[SyncIO, ExecutionContext] =
+    Resource.make(SyncIO {
+      ConstellationExecutionContext.unbounded
+    })(_ => SyncIO.unit)
 
   def run(args: List[String]): IO[ExitCode] = {
     val s = for {
@@ -104,12 +100,7 @@ object ConstellationNodeUnbounded {
       registry <- Prometheus.collectorRegistry[IO]
       metrics <- Prometheus.metricsOps[IO](registry, "org_http4s_client")
 
-      contextShiftApiClient = IO.contextShift(ConstellationExecutionContext.unbounded)
-      concurrentEffectApiClient = IO.ioConcurrentEffect(contextShiftApiClient)
-
       apiClient <- {
-        implicit val ce = concurrentEffectApiClient
-
         BlazeClientBuilder[IO](ConstellationExecutionContext.unbounded)
           .withConnectTimeout(30.seconds)
           .withMaxTotalConnections(1024)
@@ -129,13 +120,13 @@ object ConstellationNodeUnbounded {
           nodeConfig.hostName,
           sessionTokenService,
           nodeConfig.primaryKeyPair.getPublic.toId
-        )(concurrentEffectApiClient, contextShiftApiClient)
+        )
 
       cloudService <- Resource.liftF(IO.delay { CloudService[IO](cloudConfig) })
       cloudQueue <- Resource.liftF(Queue.unbounded[IO, DataToSend])
       cloudQueueInstance <- Resource.liftF(cloudService.cloudSendingQueue(cloudQueue))
 
-      node = new ConstellationNode(nodeConfig, cloudQueueInstance, signedApiClient, registry, sessionTokenService)(cs)
+      node = new ConstellationNode(nodeConfig, cloudQueueInstance, signedApiClient, registry, sessionTokenService)
 
       server <- createServer(node.dao, registry)
     } yield server